// src/utils/enums.ts
export enum UserRole {
  Lawyer = 1,
  Client = 2,
}

export enum CaseStatus {
  Open = 1,
  Closed = 2,
}

export enum TodoStatus {
  Pending = 1,
  InProgress = 2,
  Completed = 3,
  Cancelled = 4,
}

export enum DayType {
  Today = 1,
  Upcoming = 2,
  Past = 3,
}

export enum CommonFilter {
  All = -1,
}

export enum OrderDir {
  Asc = 'asc',
  Desc = 'desc',
}